+++ dvh/dvh_bokeh_models/main/roi_viewer.py
@@ -11,11 +11,14 @@ from bokeh.plotting import figure
 from bokeh.layouts import row, column
 from bokeh.models import Spacer
 from tools.utilities import clear_source_data
-import options
+from tools.io.preferences.options import load_options
 from bokeh import events
 from tools.io.database.sql_connector import DVH_SQL
 from tools.roi.formatter import get_planes_from_string
 from tools.roi.geometry import union
+
+
+options = load_options()
 
 
 class RoiViewerRoiColorTicker: