+++ dvh/tools/io/database/update.py
@@ -7,6 +7,7 @@ Created on Fri Dec 28 2018
 """
 
 from __future__ import print_function
+from future.utils import listitems
 from tools.io.database.sql_connector import DVH_SQL
 from tools.roi import geometry as roi_geom
 from tools.roi import formatter as roi_form
@@ -30,11 +31,7 @@ def centroid(study_instance_uid, roi_name):
 
     data = [str(round(v, 3)) for v in data]
 
-    DVH_SQL().update('dvhs',
-                     'centroid',
-                     ','.join(data),
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
+    update_dvhs_table(study_instance_uid, roi_name, 'centroid', ','.join(data))
 
 
 def cross_section(study_instance_uid, roi_name):
@@ -52,17 +49,8 @@ def cross_section(study_instance_uid, roi_name):
     roi = roi_form.get_planes_from_string(coordinates_string[0][0])
     area = roi_geom.cross_section(roi)
 
-    DVH_SQL().update('dvhs',
-                     'cross_section_max',
-                     area['max'],
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
-
-    DVH_SQL().update('dvhs',
-                     'cross_section_median',
-                     area['median'],
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
+    for key in ['max', 'median']:
+        update_dvhs_table(study_instance_uid, roi_name, 'cross_section_%s' % key, area[key])
 
 
 def spread(study_instance_uid, roi_name):
@@ -82,21 +70,8 @@ def spread(study_instance_uid, roi_name):
 
     data = [str(round(v/10., 3)) for v in data]
 
-    DVH_SQL().update('dvhs',
-                     'spread_x',
-                     data[0],
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
-    DVH_SQL().update('dvhs',
-                     'spread_y',
-                     data[1],
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
-    DVH_SQL().update('dvhs',
-                     'spread_z',
-                     data[2],
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
+    for i, column in enumerate(['spread_x', 'spread_y', 'spread_z']):
+        update_dvhs_table(study_instance_uid, roi_name, column, data[i])
 
 
 def min_distances(study_instance_uid, roi_name):
@@ -128,35 +103,15 @@ def min_distances(study_instance_uid, roi_name):
             dth = roi_geom.dth(data)
             dth_string = ','.join(['%.3f' % num for num in dth])
 
-            DVH_SQL().update('dvhs',
-                             'dist_to_ptv_min',
-                             round(float(np.min(data)), 2),
-                             "study_instance_uid = '%s' and roi_name = '%s'"
-                             % (study_instance_uid, roi_name))
+            data_map = {'dist_to_ptv_min': round(float(np.min(data)), 2),
+                        'dist_to_ptv_mean': round(float(np.mean(data)), 2),
+                        'dist_to_ptv_median': round(float(np.median(data)), 2),
+                        'dist_to_ptv_max': round(float(np.max(data)), 2),
+                        'dth_string': dth_string}
 
-            DVH_SQL().update('dvhs',
-                             'dist_to_ptv_mean',
-                             round(float(np.mean(data)), 2),
-                             "study_instance_uid = '%s' and roi_name = '%s'"
-                             % (study_instance_uid, roi_name))
+            for key, value in listitems(data_map):
+                update_dvhs_table(study_instance_uid, roi_name, key, value)
 
-            DVH_SQL().update('dvhs',
-                             'dist_to_ptv_median',
-                             round(float(np.median(data)), 2),
-                             "study_instance_uid = '%s' and roi_name = '%s'"
-                             % (study_instance_uid, roi_name))
-
-            DVH_SQL().update('dvhs',
-                             'dist_to_ptv_max',
-                             round(float(np.max(data)), 2),
-                             "study_instance_uid = '%s' and roi_name = '%s'"
-                             % (study_instance_uid, roi_name))
-
-            DVH_SQL().update('dvhs',
-                             'dth_string',
-                             dth_string,
-                             "study_instance_uid = '%s' and roi_name = '%s'"
-                             % (study_instance_uid, roi_name))
         except:
             print('dist_to_ptv calculation failure, skipping')
 
@@ -183,14 +138,10 @@ def treatment_volume_overlap(study_instance_uid, roi_name):
 
         ptvs = [roi_form.get_planes_from_string(ptv[0]) for ptv in ptv_coordinates_strings]
 
-        tv = roi_form.get_union(ptvs)
+        tv = roi_geom.union(ptvs)
         overlap = roi_geom.overlap_volume(oar, tv)
 
-        DVH_SQL().update('dvhs',
-                         'ptv_overlap',
-                         round(float(overlap), 2),
-                         "study_instance_uid = '%s' and roi_name = '%s'"
-                         % (study_instance_uid, roi_name))
+        update_dvhs_table(study_instance_uid, roi_name, 'ptv_overlap', round(float(overlap), 2))
 
 
 def dist_to_ptv_centroids(study_instance_uid, roi_name):
@@ -219,11 +170,7 @@ def dist_to_ptv_centroids(study_instance_uid, roi_name):
 
         data = float(np.linalg.norm(ptv_centroid - oar_centroid)) / 10.
 
-        DVH_SQL().update('dvhs',
-                         'dist_to_ptv_centroids',
-                         round(data, 3),
-                         "study_instance_uid = '%s' and roi_name = '%s'"
-                         % (study_instance_uid, roi_name))
+        update_dvhs_table(study_instance_uid, roi_name, 'dist_to_ptv_centroids', round(float(data), 3))
 
 
 def volumes(study_instance_uid, roi_name):
@@ -242,11 +189,7 @@ def volumes(study_instance_uid, roi_name):
 
     data = roi_geom.volume(roi)
 
-    DVH_SQL().update('dvhs',
-                     'volume',
-                     round(float(data), 2),
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
+    update_dvhs_table(study_instance_uid, roi_name, 'volume', round(float(data), 2))
 
 
 def surface_area(study_instance_uid, roi_name):
@@ -265,8 +208,9 @@ def surface_area(study_instance_uid, roi_name):
 
     data = roi_geom.surface_area(roi, coord_type="sets_of_points")
 
-    DVH_SQL().update('dvhs',
-                     'surface_area',
-                     round(float(data), 2),
-                     "study_instance_uid = '%s' and roi_name = '%s'"
-                     % (study_instance_uid, roi_name))
+    update_dvhs_table(study_instance_uid, roi_name, 'surface_area', round(float(data), 2))
+
+
+def update_dvhs_table(study_instance_uid, roi_name, column, value):
+    DVH_SQL().update('dvhs', column, value,
+                     "study_instance_uid = '%s' and roi_name = '%s'" % (study_instance_uid, roi_name))